requirements.txt
streamlit
tensorflowtensorflow==2.19.0
opencv-python
numpy
scipy
faiss-cpu
Pillow
werkzeug
